Move hooks to top-level
