+++ src/modals/SendModals/SendModalDeployContract.tsx
@@ -24,12 +24,12 @@ import InfoBox from '../../components/InfoBox'
 import Input from '../../components/Inputs/Input'
 import { Client } from '../../contexts/global'
 import { useSendModalContext } from '../../contexts/sendModal'
+import { useBuildTxCommonComponents, useBytecodeInputComponent } from '../../hooks/modals/SendModals/hooks'
 import useDappTxData from '../../hooks/useDappTxData'
 import { DeployContractTxData } from '../../types/transactions'
 import { isAmountWithinRange } from '../../utils/transactions'
 import AlphAmountInfoBox from './AlphAmountInfoBox'
 import BuildTxFooterButtons from './BuildTxFooterButtons'
-import { useBuildTxCommonComponents, useBytecodeInputComponent } from './hooks'
 import { ModalInputFields } from './ModalInputFields'
 import SendModal, { TxContext } from './SendModal'
 import { CheckTxProps, PartialTxData } from './types'
