+++ src/modals/SendModals/SendModalScript.tsx
@@ -22,12 +22,12 @@ import { SignExecuteScriptTxResult } from 'alephium-web3'
 import InfoBox from '../../components/InfoBox'
 import { Client } from '../../contexts/global'
 import { useSendModalContext } from '../../contexts/sendModal'
+import { useBuildTxCommonComponents, useBytecodeInputComponent } from '../../hooks/modals/SendModals/hooks'
 import useDappTxData from '../../hooks/useDappTxData'
 import { ScriptTxData } from '../../types/transactions'
 import { isAmountWithinRange } from '../../utils/transactions'
 import AlphAmountInfoBox from './AlphAmountInfoBox'
 import BuildTxFooterButtons from './BuildTxFooterButtons'
-import { useBuildTxCommonComponents, useBytecodeInputComponent } from './hooks'
 import { ModalInputFields } from './ModalInputFields'
 import SendModal, { TxContext } from './SendModal'
 import { CheckTxProps, PartialTxData } from './types'
