+++ src/modals/SendModals/SendModalTransfer.tsx
@@ -24,13 +24,13 @@ import InfoBox from '../../components/InfoBox'
 import Input from '../../components/Inputs/Input'
 import { Client } from '../../contexts/global'
 import { useSendModalContext } from '../../contexts/sendModal'
+import { useBuildTxCommonComponents } from '../../hooks/modals/SendModals/hooks'
 import useDappTxData from '../../hooks/useDappTxData'
 import { TransferTxData } from '../../types/transactions'
 import { isAddressValid } from '../../utils/addresses'
 import { isAmountWithinRange } from '../../utils/transactions'
 import AlphAmountInfoBox from './AlphAmountInfoBox'
 import BuildTxFooterButtons from './BuildTxFooterButtons'
-import { useBuildTxCommonComponents } from './hooks'
 import { ModalInputFields } from './ModalInputFields'
 import SendModal, { TxContext } from './SendModal'
 import { CheckTxProps, PartialTxData } from './types'
